requirements.txt
Flask
Flask-SQLAlchemy
psycopg2-binary
pandas
numpy
scikit-learn
spacy
sentence-transformers
pdfplumber
python-docx
nltk
tqdm
joblib
tensorflow

